# be_agent_orchestrator/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Orchestrator",
    "description": "\nYuno AI Agent Orchestration Platform \u2014 backend API.\n\nStack: FastAPI \u00b7 LangGraph \u00b7 PostgreSQL \u00b7 Redis \u00b7 Arq.\n\n**Auth.** All non-public endpoints require an `Authorization: Bearer <access_token>`\nheader. Obtain tokens via `POST /api/v1/auth/login` and rotate via\n`POST /api/v1/auth/refresh`. The access token is valid for 12 hours; the\nrefresh token for 30 days.\n\n**Error envelope.** Every non-2xx response is `{\"error\": {\"code\", \"message\", \"details\"}}`.\nSwitch on `code` in the frontend (it's stable); render `message` to the user;\ninspect `details` for structured per-error data.\n",
    "version": "0.1.0",
    "contact": {
      "name": "Yuno AI",
      "email": "engineering@yuno.ai"
    },
    "license": {
      "name": "Proprietary"
    },
    "x-environment": "local"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local dev"
    },
    {
      "url": "http://api:3001",
      "description": "Docker compose internal"
    }
  ],
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/agents/": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List agents inside a workspace",
        "description": "Returns every agent in the workspace, newest first. Light shape: id, name, description, created_at, created_by. Use `GET /workspaces/{ws_id}/agents/{agent_id}` for the full config.",
        "operationId": "agents_get_list_in_workspace",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Workspace UUID.",
              "title": "Workspace Id"
            },
            "description": "Workspace UUID."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_AgentSummary_"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error on request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Create an agent in a workspace",
        "description": "Creates an agent (workflow node). The five mandatory fields are the minimum LangGraph needs to materialize the node at runtime via `create_react_agent(model, tools, prompt=...)`:\n\n- `name` \u2014 display name\n- `role` \u2014 short role label (used in conversation routing)\n- `system_prompt` \u2014 the LLM's system message\n- `model_provider` \u2014 LLM provider key (must be a configured one)\n- `model_name` \u2014 provider-specific model id\n\nAll other fields (description, sampling params, memory/guardrails/limits/interaction-rules/skills configs) default to sensible values and can be edited later via the agent editor.",
        "operationId": "agents_post_create",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Workspace UUID.",
              "title": "Workspace Id"
            },
            "description": "Workspace UUID."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error on request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/capabilities": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get supported agent-creation configs",
        "description": "Returns everything the agent-builder UI needs to render the create-agent form without hard-coding anything:\n\n- `providers` \u2014 every LLM provider the server knows about, with a curated recommended-model list, the **per-user** `is_configured` flag (true iff either the server has a global key for it OR this user has saved one in their LLM credentials vault), and per-agent config fields with bounds.\n- `fields.mandatory` / `fields.optional` \u2014 every `CreateAgentRequest` field split by required-ness, with type/length/range constraints.\n- `tools` \u2014 built-in tools available to attach to agents (empty in v1 until the tool registry is wired).\n- `tool_entry_shape` \u2014 the expected shape of each item when attaching tools to an agent's `skills_config.tools` array.\n- `provider_credentials_shape` \u2014 schema for the optional per-agent BYOK field on agent creation.",
        "operationId": "agents_get_capabilities",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCapabilities"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error on request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentCapabilities": {
        "properties": {
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderCapability"
            },
            "type": "array",
            "title": "Providers"
          },
          "fields": {
            "$ref": "#/components/schemas/AgentFieldCatalog"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/ToolCapability"
            },
            "type": "array",
            "title": "Tools",
            "description": "Built-in tools available to attach to agents. May be empty in v1 until the tool registry is wired."
          },
          "tool_entry_shape": {
            "$ref": "#/components/schemas/ToolEntryShape",
            "description": "Shape of each entry to add when attaching tools on an agent's skills_config."
          },
          "provider_credentials_shape": {
            "$ref": "#/components/schemas/ProviderCredentialsShape",
            "description": "Schema for the optional `provider_credentials` form field on create."
          }
        },
        "type": "object",
        "required": [
          "providers",
          "fields",
          "provider_credentials_shape"
        ],
        "title": "AgentCapabilities",
        "description": "Everything needed to render the agent-creation form.",
        "examples": [
          {
            "fields": {
              "mandatory": [
                {
                  "description": "Display name.",
                  "key": "name",
                  "max_length": 255,
                  "min_length": 1,
                  "required": true,
                  "type": "string"
                }
              ],
              "optional": []
            },
            "providers": [
              {
                "config_fields": [
                  {
                    "default": 0.7,
                    "description": "Sampling temperature.",
                    "key": "temperature",
                    "max": 2.0,
                    "min": 0.0,
                    "required": false,
                    "type": "float"
                  }
                ],
                "default_model": "gpt-4o-mini",
                "display_name": "OpenAI",
                "is_configured": true,
                "key": "openai",
                "models": [
                  {
                    "context_window": 128000,
                    "display_name": "GPT-4o mini",
                    "is_default": true,
                    "name": "gpt-4o-mini",
                    "supports_streaming": true,
                    "supports_tools": true
                  }
                ]
              }
            ],
            "tool_entry_shape": {
              "name_field": "name",
              "options_field": "options"
            },
            "tools": []
          }
        ]
      },
      "AgentDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Short role label, e.g. `triage`, `researcher`."
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "archived"
            ],
            "title": "Status"
          },
          "model_provider": {
            "type": "string",
            "title": "Model Provider",
            "description": "`openai` | `gemini` | `groq` | \u2026"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "temperature": {
            "type": "number",
            "title": "Temperature"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "top_p": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top P"
          },
          "memory_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Memory Config"
          },
          "schedule_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Schedule Config"
          },
          "guardrails_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Guardrails Config"
          },
          "interaction_rules": {
            "additionalProperties": true,
            "type": "object",
            "title": "Interaction Rules"
          },
          "limits_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Limits Config"
          },
          "skills_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Skills Config"
          },
          "provider_credentials_configured": {
            "type": "boolean",
            "title": "Provider Credentials Configured",
            "description": "True iff the agent carries its own provider credentials. The actual key is never returned by any endpoint.",
            "default": false
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "name",
          "slug",
          "role",
          "system_prompt",
          "status",
          "model_provider",
          "model_name",
          "temperature",
          "created_at",
          "updated_at"
        ],
        "title": "AgentDetail",
        "description": "Full agent configuration. Returned by the detail endpoint.\n\nThe JSONB blobs (memory/schedule/guardrails/interaction_rules/limits/\nskills) are returned as opaque dicts here \u2014 frontend renders them via\nthe agent-editor module. Backend doesn't impose a schema on these\nfields yet; that comes when we wire the runtime executor."
      },
      "AgentFieldCatalog": {
        "properties": {
          "mandatory": {
            "items": {
              "$ref": "#/components/schemas/FieldSpec"
            },
            "type": "array",
            "title": "Mandatory"
          },
          "optional": {
            "items": {
              "$ref": "#/components/schemas/FieldSpec"
            },
            "type": "array",
            "title": "Optional"
          }
        },
        "type": "object",
        "required": [
          "mandatory",
          "optional"
        ],
        "title": "AgentFieldCatalog"
      },
      "AgentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By",
            "description": "User who authored the agent. Null if that user was deleted."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "AgentSummary",
        "description": "Light shape for the list endpoint \u2014 no JSONB configs (kilobytes+ per row).",
        "examples": [
          {
            "created_at": "2026-05-23T18:42:13.812000Z",
            "created_by": "01HZX9F4Y0N0N5J5R0Q6Z3M8AB",
            "description": "Classifies inbound questions and hands off to a specialist.",
            "id": "01HZX9F4Y0N0N5J5R0Q6Z3M8B2",
            "name": "Support Router"
          }
        ]
      },
...
      "CreateAgentRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Human-readable display name shown in the workflow builder."
          },
          "role": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Role",
            "description": "Short role label (e.g. `triage`, `researcher`, `chart_generator`)."
          },
          "system_prompt": {
            "type": "string",
            "minLength": 1,
            "title": "System Prompt",
            "description": "System prompt the LLM receives. Becomes `prompt=` in `create_react_agent`."
          },
          "model_provider": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Model Provider",
            "description": "LLM provider key. Must match a configured provider (`openai`, `gemini`, `groq`)."
          },
          "model_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Model Name",
            "description": "Provider-specific model identifier, e.g. `gpt-4o-mini`, `gemini-2.0-flash`, `llama-3.3-70b-versatile`."
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Free-text description shown in the agent picker / hover card."
          },
          "temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 0.7
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 128000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "top_p": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top P"
          },
          "memory_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Memory Config"
          },
          "schedule_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Schedule Config"
          },
          "guardrails_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Guardrails Config"
          },
          "interaction_rules": {
            "additionalProperties": true,
            "type": "object",
            "title": "Interaction Rules"
          },
          "limits_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Limits Config"
          },
          "skills_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Skills Config"
          },
          "provider_credentials": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderCredentialsConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-agent LLM provider credentials. When omitted, the agent uses the server-level API key for `model_provider` from `.env`. **Required** when the provider is not globally configured."
          }
        },
        "type": "object",
        "required": [
          "name",
          "role",
          "system_prompt",
          "model_provider",
          "model_name"
        ],
        "title": "CreateAgentRequest",
        "examples": [
          {
            "description": "Classifies inbound tickets and dispatches to specialists.",
            "model_name": "gpt-4o-mini",
            "model_provider": "openai",
            "name": "Support Router",
            "role": "router",
            "system_prompt": "You are the front-line router for inbound customer messages. Classify each message into one of: billing, technical, sales. Hand off to the appropriate specialist agent.",
            "temperature": 0.2
          }
        ]
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable upper-snake error code. Switch on this in the UI."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable summary; safe to surface to the user."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "description": "Structured per-error data (validation errors, conflict keys, etc)."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "The `error` field of every non-2xx response.\n\nMirrors what ``app.core.errors.register_exception_handlers`` emits, so\nthe OpenAPI schema matches the real wire format byte-for-byte.",
        "examples": [
          {
            "code": "AGENT_NOT_FOUND",
            "details": {
              "agent_id": "01HZ..."
            },
            "message": "Agent not found"
          }
        ]
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "Top-level error response shape returned by every error path.",
        "examples": [
          {
            "error": {
              "code": "VALIDATION_ERROR",
              "details": {
                "errors": [
                  {
                    "loc": [
                      "body",
                      "email"
                    ],
                    "msg": "field required"
                  }
                ]
              },
              "message": "Invalid request payload"
            }
          }
        ]
      },
...
      "PageResponse_AgentSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PageResponse[AgentSummary]"
      },
...
      "ProviderCapability": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Use this as `model_provider` when creating an agent."
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "is_configured": {
            "type": "boolean",
            "title": "Is Configured",
            "description": "True iff the server has an API key for this provider. Agents using a non-configured provider cannot be executed."
          },
          "default_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Model",
            "description": "The model_name we recommend by default. Null if no models are catalogued."
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelCatalogEntry"
            },
            "type": "array",
            "title": "Models"
          },
          "config_fields": {
            "items": {
              "$ref": "#/components/schemas/ProviderConfigField"
            },
            "type": "array",
            "title": "Config Fields",
            "description": "Per-agent LLM config knobs (temperature, max_tokens, top_p) with bounds."
          }
        },
        "type": "object",
        "required": [
          "key",
          "display_name",
          "is_configured",
          "models",
          "config_fields"
        ],
        "title": "ProviderCapability"
      },
...
      "ProviderCredentialsConfig": {
        "properties": {
          "api_key": {
            "type": "string",
            "minLength": 1,
            "title": "Api Key",
            "description": "Provider API key (e.g. `sk-...`, `gsk_...`, `AIza...`)."
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "Custom API base URL. Useful for proxies (OpenAI-compatible endpoints, Groq mirrors)."
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization",
            "description": "OpenAI organization id, if applicable."
          }
        },
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "ProviderCredentialsConfig",
        "description": "Per-agent override of the server-level LLM provider credentials.\n\nOptional. When set, ``api_key`` overrides ``settings.<provider>_api_key``\nfor this agent's runs. ``base_url`` + ``organization`` are honored where\nthe provider's SDK supports them (OpenAI; Groq supports `base_url`).\n\nSecurity:\n- The api_key is **stored as-is in JSONB** (Postgres TDE only). Treat\n  Postgres read access as equivalent to having the key.\n- The api_key is **never returned in any GET response** \u2014 only a\n  ``provider_credentials_configured: bool`` flag.\n- Whole ``provider_credentials`` subtree is redacted in HTTP access logs.",
        "examples": [
          {
            "api_key": "sk-proj-..."
          },
          {
            "api_key": "gsk_...",
            "base_url": "https://api.groq.com/openai/v1"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste your access token (without the `Bearer ` prefix)."
      }
    }
  },
...
}